+++ source/modulo_components/test/cpp/include/test_modulo_components/component_public_interfaces.hpp
@@ -35,8 +35,10 @@ public:
   using ComponentInterface<NodeT>::empty_services_;
   using ComponentInterface<NodeT>::string_services_;
   using ComponentInterface<NodeT>::add_tf_broadcaster;
+  using ComponentInterface<NodeT>::add_static_tf_broadcaster;
   using ComponentInterface<NodeT>::add_tf_listener;
   using ComponentInterface<NodeT>::send_transform;
+  using ComponentInterface<NodeT>::send_static_transform;
   using ComponentInterface<NodeT>::lookup_transform;
   using ComponentInterface<NodeT>::raise_error;
   using ComponentInterface<NodeT>::get_qos;